Create End-to-End-Machine-Learning-Project
An End to End machine learning model follows the complete lifecycle of a machine learning model which means first to collect data, then prepare the data according to the patterns found in the dataset, then train and evaluate the model and finally deploy the model. So an end-to-end machine learning project means presenting a model in the form of an interactive application that runs the model on the dataset and produces output in real-time according to the user input. In this article, I will take you through an end-to-end machine-learning project using Python.

End-to-End Machine Learning Project
An end-to-end machine learning project means creating an interactive application that runs our trained machine learning model and gives output according to the user input. It follows the complete lifecycle of a machine-learning model. Below are the steps that you need to follow while creating an end-to-end application for your model:

Data access and collection
Data preparation and exploration
Model training and evaluation
Model deployment
Creating an end-to-end machine learning application is essential to show most of your skills in a single project. In the section below, I will take you through how to create an end-to-end machine-learning application using Python.